rt-thread: Add periodic task creation function
RT-Thread provides only with one shot tasks and it's up to the
user to create the timer that handles the priodicity of a task.

Create a "rt_thread_create_periodic()" function that allows the
user to specify the periodicity of the task.

The function sets as a default callback the rt_periodic_task_f()
function which enters in an infinite loop and everytime supsends
itself after calling the function combined to the task. A timer
takes care of resuming the task at each period.

Signed-off-by: Andi Shyti <andi.shyti@studenti.polito.it>