Write/read GPIO tasks using mailbox
The SPS needs a set of tasks to write and read on GPIO, in order
to control the status of the target machines.

As there is no real hardware -but simulated in QEMU- it is needed
to share the status of the GPIO between tasks.
One task owned by the SPS writes on GPIO. A different periodic
task emulates an external HW to simulated each of the target
devices connected to the SPS.

For this, the mailbox functionalities provided by RT-Thread are
used. The first task sends a mail to the mailbox, a second task
can retrieve the mail.

Only 32 bits of data are allowed per mail, so variable gpio and
status are merge into one variable of 32 bits.

As testing, two aperiodic task send a mail to the mailbox. One
periodic task retrieve each mail.

Signed-off-by: Andi Shyti <andi.shyti@studenti.polito.it>
Signed-off-by: Daniel Guarecuco <daniel.guarecuco@studenti.polito.it>