sps: add SPS Smart Power Switch implementation
The SPS is a Smart Power Switch that interfaces one one end with a user
and with a given number of targets on the other end. Throught the SPS
the user can control the power status of targets deciding whether he
wants to shut them down or power them on.

Add the main implementation of it. The SPS is currently handled by a
main() function which triggers its initialization and startup. The
library returns an SPS reference.

The SPS checks for the status of the target machines and wait for the
user to send power on/off commands.

Signed-off-by: Daniel Guarecuco <daniel.guarecuco@studenti.polito.it>
Signed-off-by: Andi Shyti <andi.shyti@studenti.polito.it>