target: Add the simulation of target machines
Targets simulate the behavior of the machines connected in array
to the SPS. We can add as many as we want, but for now, for readability
reasons we have added only four.

The SPS can power on and off targets by raising or dropping a GPIO
line that is connected to a master power switch that provides power
to all the targets. It's the user that tells SPS which targets to
power on or off.

Each target is connected to the network and the SPS conitnuosly
monitors their status by pinging them. If a target went frozen,
the SPS notifies the host through an irq.

The Simulation process puts targets in three possible states:

 - ON: the user has requested to power on a device
 - OFF: the user has requested to power off a device
 - FROZEN: the machine was on but it froze. The decision to freeze
   is taken with a given probability during the targets operation
   (when in ON status). In order to recover the user needs to first
   power it off and then, if necessary, power it on again.

Signed-off-by: Daniel Guarecuco <daniel.guarecuco@studenti.polito.it>
Reviewed-by: Andi Shyti <andi.shyti@studenti.polito.it>