main: Add stress tests
We want to push the system to the limit in order to check when:

 1. we run out of memory and overflow the stack
 2. we start missing deadlines
 3. we start using the CPU next to 100%

For doing that we assign some random operations to the targets,
e.g. matrix-matrix multiplicatin, executed enough time to require
more resources in scheduling time and memory.

Thanks to the CPU usage we can also monitor how much the CPU is
used.

(For not abusing of the stack memory availability the
matrix-matrix multiplication is exetuded N times instead of
increasing the size of the matrix)

The paramenters we change are:

 1. Stack size
 2. number of threads
 3. matrix size
 4. number of times we execute the multiplication

Signed-off-by: Andi Shyti <andi.shyti@studenti.polito.it>
Signed-off-by: Daniel Guarecuco <daniel.guarecuco@studenti.polito.it>